handled errors and added static files to git

+++ cakebakery/accounts/views.py
@@ -2,7 +2,7 @@ from django.shortcuts import render,reverse,redirect
 from django.contrib.auth.models import User
 from django.contrib.auth.forms import UserCreationForm,AuthenticationForm
 from django.contrib.auth import login,logout,authenticate
-from django.contrib import messages
+from django.contrib import messages as mssg
 from products.models import cake_list,orders,messages
 from .forms import SignUpForm
 from products.forms import make_order_form,message_form
@@ -32,10 +32,13 @@ def sign_up(request):
     context= dict()
     context["form"] = form
     if request.method == "POST":
-        if form.is_valid: 
+        if form.is_valid(): 
             user=form.save()
             login(request,user)      
             return redirect(reverse('accounts:homepage'))
+        else:
+            mssg.info(request,form.errors)
+            return render(request,'accounts/sign_up.html',context)
     return render(request,'accounts/sign_up.html',context)
 
 def log_in(request):
@@ -43,7 +46,7 @@ def log_in(request):
     context= dict()
     context["form"] = form
     if request.method == "POST":
-        if form.is_valid:
+        if form.is_valid():
             username = request.POST.get('username')
             password = request.POST.get('password')
             user = authenticate(username = username,password=password)
@@ -52,6 +55,8 @@ def log_in(request):
                 return redirect(reverse('admin:dashboard'))
             if user:    
                 return redirect(reverse('accounts:homepage'))
+        else:
+            return render(request,'accounts/log_in.html',context)    
     return render(request,'accounts/log_in.html',context)
 
 def log_out(request):
@@ -65,6 +70,14 @@ def my_orders(request):
     context['orders'] = myorders
     return render(request,'accounts/my_orders.html',context)
 
+def cancel_order(request,id):
+    order = orders.objects.get(id = id)
+    order.order_status = 'Cancelled'
+    order.save()
+    mssg.info(request,'Order cancelled successfully. Refund will be credited to your bank account within 3-4 working days')
+    return redirect(reverse('accounts:my-orders'))
+
+
 def messages(request):
     context = dict()
     form = message_form(request.POST or None)
@@ -72,6 +85,7 @@ def messages(request):
     if request.method == "POST":
         if form.is_valid: 
             order=form.save()
+            mssg.info(request,'Thank you for filling the form. Will get back to you soon')
     return render(request,'accounts/messages.html',context)
 
 def checkout(request,id):
@@ -95,5 +109,12 @@ def landing_page(request):
 
 def payment_cancelled(request):
     cancel_order = orders.objects.latest('id')
-    cancel_order.delete()
+    if cancel_order.status == 'Pending':
+        cancel_order.delete()
     return redirect(reverse('accounts:homepage'))
+
+def payment_successfull(request):
+    success_order = orders.objects.latest('id')
+    success_order.status = 'Success'
+    success_order.save()
+    return redirect(reverse('accounts:my-orders'))
